Change from svg to Chakra ui Icon
